Improve rem2ics out.ics

+++ remind.py
@@ -719,7 +719,7 @@ def rem2ics() -> None:
     )
     args = parser.parse_args()
 
-    if args.infile and args.infile != "-" and not isfile(args.infile):
+    if args.infile and args.infile != "-" and not isfile(args.infile) and not args.outfile:
         args.outfile = open(args.infile, "w", encoding="utf-8")
         args.infile = None
